implementing SignalR

+++ Proz_DesktopApplication/Sub-UserControls/settings.xaml.cs
@@ -4,9 +4,11 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 using System.Windows.Media;
+using Microsoft.AspNetCore.SignalR.Client;
 using Microsoft.Extensions.DependencyInjection;
 using ModernMessageBoxLib;
 using Proz_DesktopApplication.API;
+using Proz_DesktopApplication.HelperServices;
 
 namespace Proz_DesktopApplication.Sub_UserControls
 {
@@ -18,6 +20,7 @@ namespace Proz_DesktopApplication.Sub_UserControls
         public AdminAPIEndpointsDefinitions adminAPIEndpointsDefinitions;
         private bool _hasLoadedOnce = false;
         string? CompanyName;
+        private MainHubService MainHubConnection;
         public settings()
         {
             InitializeComponent();
@@ -25,6 +28,7 @@ namespace Proz_DesktopApplication.Sub_UserControls
 
             this.Loaded += UserControl_Loaded;
             this.Loaded += LoadingData;
+   
         }
 
         private void UserControl_Loaded(object sender, RoutedEventArgs e)
@@ -66,6 +70,7 @@ namespace Proz_DesktopApplication.Sub_UserControls
             _services = Services1 ?? throw new InvalidOperationException("Services1 is null");
             _authApi = AuthApi1 ?? throw new InvalidOperationException("AuthApi1 is null");
             adminAPIEndpointsDefinitions = _AdminAPIEndpointsDefinitions1 ?? throw new InvalidOperationException("Services1 is null");
+            MainHubConnection = MainHub ?? throw new InvalidOperationException("MainHUB is null");
         }
 
         private async void LoadingData(object sender, RoutedEventArgs e)
@@ -621,17 +626,52 @@ namespace Proz_DesktopApplication.Sub_UserControls
                 return;
             }
         }
-        private void LogOutButton_Click(object sender, RoutedEventArgs e)
+        private async void LogOutButton_Click(object sender, RoutedEventArgs e)
         {
-            
-            var signInWindow = _services.GetRequiredService<SigninWindow>();
-            
-            TokenStorage.DeleteTokens();
-            Application.Current.MainWindow = signInWindow;
-            signInWindow.Show();
-            var mainWindow = Window.GetWindow(this) as MainDashboardWindow;
 
-            mainWindow?.Close();
+            //var signInWindow = _services.GetRequiredService<SigninWindow>();
+
+            //TokenStorage.DeleteTokens();
+            //Application.Current.MainWindow = signInWindow;
+            //signInWindow.Show();
+            //var mainWindow = Window.GetWindow(this) as MainDashboardWindow;
+
+            //mainWindow?.Close();
+
+            var signInWindow = _services.GetRequiredService<SigninWindow>();
+            string DeviceToken = TokenStorage.GetOrCreateDeviceToken();
+            var tokens = TokenStorage.LoadTokens();
+            try
+            {
+                var request = new LogoutRequest { DeviceToken = DeviceToken, RefreshToken = tokens.Value.refreshToken };
+                var win = new IndeterminateProgressWindow("Logging out, please wait...");
+                win.Show();
+                var response = await _authApi.Logout(request);
+                win.Close();
+
+
+                if (response.IsSuccessStatusCode)
+                {
+                    TokenStorage.DeleteTokens();
+                    Application.Current.MainWindow = signInWindow;
+                    signInWindow.Show();
+                    var mainWindow = Window.GetWindow(this) as MainDashboardWindow;
+                   await MainHubConnection.Connection.StopAsync();
+                    mainWindow?.Close();
+
+                }
+                else
+                {
+                    var msg = new ModernMessageBox($"Couldn't log out, make sure the server is on.", "Something went wrong!", ModernMessageboxIcons.Error, "OK");
+                    msg.ShowDialog();
+                }
+            }
+            catch (Exception ex)
+            {
+                var msg = new ModernMessageBox($"Couldn't reach the server.", "Something went wrong!", ModernMessageboxIcons.Error, "OK");
+                msg.ShowDialog();
+
+            }
