+++ Proz_DesktopApplication/Sub-UserControls/settingsAdmin.xaml.cs
@@ -1,6 +1,8 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.AspNetCore.SignalR.Client;
+using Microsoft.Extensions.DependencyInjection;
 using ModernMessageBoxLib;
 using Proz_DesktopApplication.API;
+using Proz_DesktopApplication.HelperServices;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -29,11 +31,13 @@ namespace Proz_DesktopApplication.Sub_UserControls
         public AdminAPIEndpointsDefinitions adminAPIEndpointsDefinitions;
         private bool _hasLoadedOnce = false;
         string? CompanyName;
+        private MainHubService MainHubConnection;
         public settingsAdmin()
         {
             InitializeComponent();
             this.Loaded += settings_Loaded;
             this.Loaded += LoadingData;
+          
         }
            private void settings_Loaded(object sender, RoutedEventArgs e)
            {
@@ -74,6 +78,7 @@ namespace Proz_DesktopApplication.Sub_UserControls
             _services = Services1 ?? throw new InvalidOperationException("Services1 is null");
             _authApi = AuthApi1 ?? throw new InvalidOperationException("AuthApi1 is null");
             adminAPIEndpointsDefinitions = _AdminAPIEndpointsDefinitions1 ?? throw new InvalidOperationException("Services1 is null");
+            MainHubConnection = MainHub ?? throw new InvalidOperationException("MainHUB is null");
         }
         private async void LoadingData(object sender, RoutedEventArgs e)
         {
@@ -629,17 +634,52 @@ namespace Proz_DesktopApplication.Sub_UserControls
             }
         }
 
-        private void LogOutButton_Click(object sender, RoutedEventArgs e)
+        private async void LogOutButton_Click(object sender, RoutedEventArgs e)
         {
 
+            //var signInWindow = _services.GetRequiredService<SigninWindow>();
+
+            //TokenStorage.DeleteTokens();
+            //Application.Current.MainWindow = signInWindow;
+            //signInWindow.Show();
+            //var mainWindow = Window.GetWindow(this) as MainDashboardWindow;
+
+            //mainWindow?.Close();
+
             var signInWindow = _services.GetRequiredService<SigninWindow>();
+            string DeviceToken = TokenStorage.GetOrCreateDeviceToken();
+            var tokens = TokenStorage.LoadTokens();
+            try
+            {
+                var request = new LogoutRequest { DeviceToken = DeviceToken, RefreshToken = tokens.Value.refreshToken };
+                var win = new IndeterminateProgressWindow("Logging out, please wait...");
+                win.Show();
+                var response = await _authApi.Logout(request);
+                win.Close();
 
-            TokenStorage.DeleteTokens();
-            Application.Current.MainWindow = signInWindow;
-            signInWindow.Show();
-            var mainWindow = Window.GetWindow(this) as MainDashboardWindow;
 
-            mainWindow?.Close();
+                if (response.IsSuccessStatusCode)
+                {
+                    TokenStorage.DeleteTokens();
+                    Application.Current.MainWindow = signInWindow;
+                    signInWindow.Show();
+                    var mainWindow = Window.GetWindow(this) as MainDashboardWindow;
+                    await MainHubConnection.Connection.StopAsync();
+                    mainWindow?.Close();
+
+                }
+                else
+                {
+                    var msg = new ModernMessageBox($"Couldn't log out, make sure the server is on.", "Something went wrong!", ModernMessageboxIcons.Error, "OK");
+                    msg.ShowDialog();
+                }
+            }
+            catch (Exception ex)
+            {
+                var msg = new ModernMessageBox($"Couldn't reach the server.", "Something went wrong!", ModernMessageboxIcons.Error, "OK");
+                msg.ShowDialog();
+
+            }
